stdx

Module alloc

Unstable Memory allocation APIs

Structs

AllocError

The AllocError error indicates an allocation failure that may be due to resource exhaustion or to something wrong when combining the given input arguments with this allocator.

Global

The global memory allocator.

Traits

Allocator

An implementation of Allocator can allocate, grow, shrink, and deallocate arbitrary blocks of data described via Layout.

Functions

non_null_empty_slice

Creates a dangling NonNull<[u8]> slice of size 0 with the specified layout alignment.