stdx::anymap

Type Alias AnyMap

pub type AnyMap = Map<dyn Any>;

The most common type of Map: just using Any; [Map]<dyn [Any]>.

Why is this a separate type alias rather than a default value for Map<A>? Map::default() doesn’t seem to be happy to infer that it should go with the default value. It’s a bit sad, really. Ah well, I guess this approach will do.

Aliased Type

pub struct AnyMap {
    raw: HashMap<TypeId, Box<dyn Any>, BuildHasherDefault<TypeIdHasher>>,
}

Fields
raw: HashMap<TypeId, Box<dyn Any>, BuildHasherDefault<TypeIdHasher>>