Macro stdx::always

macro_rules! always {
    ($cond:expr) => { ... };
    ($cond:expr, $fmt:literal $($arg:tt)*) => { ... };
}

Asserts that the condition is always true and returns its actual value.

If the condition is true does nothing and and evaluates to true.

If the condition is false:

panics if force feature or debug_assertions are enabled,

logs an error if log feature is enabled,

evaluates to false.

Accepts format! style arguments.