stdx

Macro impl_from

macro_rules! impl_from {
    (
        impl<$lifetime:lifetime>
        $source:ident $(<$source_lifetime:lifetime>)?
        { $($source_variant:ident $(=> $target_variant:ident)?),* $(,)? }
        for $target:ident<$target_lifetime:lifetime>
    ) => { ... };
    (
        $source:ident
        { $($source_variant:ident $(=> $target_variant:ident)?),* $(,)? }
        for $target:ident
    ) => { ... };
    (@map_enum_variant $target:ident, $value:ident, $variant:ident) => { ... };
    (
        @map_enum_variant $target:ident, $value:ident,
        $source_variant:ident => $target_variant:ident
    ) => { ... };
    (
        impl<$lifetime:lifetime>
        $(
            $variant:ident $(<$variant_lifetime:lifetime>)?
            $(($($sub_variant:ident $(<$sub_variant_lifetime:lifetime>)?),*))?
        ),*
        for $enum:ident<$enum_lifetime:lifetime>
    ) => { ... };
    ($($variant:ident $(($($sub_variant:ident),*))?),* for $enum:ident) => { ... };
    ($($variant:ident$(<$V:ident>)?),* for $enum:ident) => { ... };
}

Generates From impls that wrap values in, or map variants between, enums.

Enum mappings with Source => Target rename the variant and convert its payload with .into().

Examples

impl_from!(Struct, Union, Enum for Adt);
impl_from!(impl<'db> Item, InternedItem<'db> for ItemId<'db>);
impl_from!(AssocItem { Function, Const, TypeAlias } for Item);
impl_from!(AdtId { StructId => Struct, EnumId => Enum } for Adt);