stdx
Module thin_vec

Re-exports

pub use crate::thin_vec_with_header_struct_ as thin_vec_with_header_struct;

Structs

EmptyOptimizedThinVec

A ThinVec that requires no allocation for the empty case.

ThinVec

ThinVecWithHeader

A type that is functionally equivalent to (Header, Box<[Item]>), but all data is stored in one heap allocation and the pointer is thin, so the whole thing’s size is like a pointer.

Traits

TrustedLen

Safety