stdx

Module variance

This is a copy of std::marker::variance.

Modules

sealed 🔒

Macros

first_token 🔒

phantom_lifetime 🔒

phantom_type 🔒

Structs

PhantomContravariant

Zero-sized type used to mark a type parameter as contravariant.

PhantomContravariantLifetime

Zero-sized type used to mark a lifetime as contravariant.

PhantomCovariant

Zero-sized type used to mark a type parameter as covariant.

PhantomCovariantLifetime

Zero-sized type used to mark a lifetime as covariant.

PhantomInvariant

Zero-sized type used to mark a type parameter as invariant.

PhantomInvariantLifetime

Zero-sized type used to mark a lifetime as invariant.

Traits

Variance

A marker trait for phantom variance types.

Functions

variance

Construct a variance marker; equivalent to Default::default.