Crate stdx

Missing batteries for standard libraries.

Re-exports

pub use itertools;

Modules

anymap

This file is a port of only the necessary features from https://github.com/chris-morgan/anymap version 1.0.0-beta.2 for use within rust-analyzer. Copyright © 2014–2022 Chris Morgan. COPYING: https://github.com/chris-morgan/anymap/blob/master/COPYING Note that the license is changed from Blue Oak Model 1.0.0 or MIT or Apache-2.0 to MIT OR Apache-2.0

non_empty_vec

See NonEmptyVec.

panic_context

A micro-crate to enhance panic messages with context info.

process

Read both stdout and stderr of child without deadlocks.

rand

We don’t use rand, as that’s too many things for us.

thin_vec

thread

A utility module for working with threads that automatically joins threads upon drop and abstracts over operating system quality of service (QoS) APIs through the concept of a “thread intent”.

Macros

always

Asserts that the condition is always true and returns its actual value.

format_to

Appends formatted string to a String.

format_to_acc

Appends formatted string to a String and returns the String.

impl_from

Generates From impls for Enum E { Foo(Foo), Bar(Bar) } enums

never

Asserts that the condition is never true and returns its actual value.

Structs

JodChild

A std::process::Child wrapper that will kill the child on drop.

Traits

TupleExt

Functions

char_has_case

defer

equal_range_by

hash_once

is_ci

is_upper_snake_case

iter_eq_by

print_backtrace

Prints backtrace to stderr, useful for debugging.

replace

slice_tails

Returns all final segments of the argument, longest first.

timeit

to_camel_case

to_lower_snake_case

to_upper_snake_case

trim_indent