Crate stdx

Missing batteries for standard libraries.

Modules

panic_context
A micro-crate to enhance panic messages with context info.

process
Read both stdout and stderr of child without deadlocks.

Macros

always
Asserts that the condition is always true and returns its actual value.

eprintln

format_to
Appends formatted string to a String.

impl_from
Generates From impls for Enum E { Foo(Foo), Bar(Bar) } enums

never
Asserts that the condition is never true and returns its actual value.

Structs

JodChild

Functions

defer

equal_range_by

is_ci

print_backtrace

replace

timeit

to_lower_snake_case

to_upper_snake_case

trim_indent